pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hbam"
version = "0.1.0"
description = "High Burden Aging Muscle (HBAM) multi-omics integration pipeline"
requires-python = ">=3.10"
dependencies = [
    "mudata>=0.2.0",
    "anndata>=0.10.0",
    "scanpy>=1.9.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "gseapy>=1.0.0",
    "pybiomart>=0.2.0",
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "loguru>=0.7.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
mofa = ["mofapy2>=0.7.0"]
dev = ["pytest>=7.0", "pytest-cov>=4.0", "ruff>=0.1.0"]
all = ["hbam[mofa,dev]"]

[project.scripts]
hbam = "hbam.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"
